rota_inteligente_fundamentos_da_IA/requirements.txt
pandas
folium
openrouteservice
scikit-learn
python-dotenv
